third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/form-action-resubmission-window-reload-from-parent-expected.txt

CONSOLE ERROR: Refused to send form data to 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/form-action-none.html' because it violates the following Content Security Policy directive: "form-action 'none'".

This is a testharness.js-based test.
PASS The form resubmission should be blocked by the new window's CSP.
Harness: the test ran to completion.
